requirements.txt
babel
iso8601
lockfile>=0.9.1
netaddr
oslo.config
oslotest
pandas<0.14  # Until https://github.com/pydata/pandas/issues/7284 is fixed
pecan
posix_ipc
python-swiftclient
requests>=1.1
six
sqlalchemy
stevedore
tooz
voluptuous
werkzeug
